Move site providers to providers
